+++ src/winston-silent-logger.js
@@ -1,0 +1,6 @@
+module.exports = require('winston').createLogger({
+	level: 'info',
+	format: winston.format.simple(),
+	silent: true,
+	transports: [ new (winston.transports.Console)({ silent: true }) ]
+});